refactor: node version of billiards
